refactor: apply review round 1 feedback
Round 1 focus: reusability, modularity, consistency.

- R1.1 Extract resolve_team_id() into _common/lib.sh; remove 4 copies in
  schedule-followup / cancel-followup / list-my-followups / watch-for-event.
- R1.2 Extract Google userinfo fetch into utils/google-userinfo.utils.ts,
  called from both the headless OAuth completion path and the gemini-cli
  workspace helper's captureFromFile. Removes duplicated try/catch + parse.
- R1.3 Centralize Google OAuth constants (client_id, redirect_uri, auth base,
  cloud function URL, refresh path, userinfo endpoint, default scopes) in
  config/oauth.config.ts. Previously duplicated between credentials.controller
  and gemini-cli-workspace.helper.
- R1.4 Split the 448-line credentials.controller into two focused modules:
  credentials.controller.ts keeps CRUD + gemini-cli import/clear endpoints;
  google-oauth.controller.ts owns the headless start/complete flow. Move the
  corresponding tests to google-oauth.controller.test.ts.

All 93 jest tests and 31 shell tests pass; typecheck clean.
