+++ backend/src/controllers/credentials/credentials.controller.ts
@@ -246,202 +246,5 @@ export async function clearGeminiCliExtensionFile(
   }
 }
 
-// ============================================================================
-// Headless OAuth flow — for remote users (orchestrator over Slack, etc.)
-// ============================================================================
-
-/**
- * OAuth client + endpoint details lifted from the Gemini CLI Workspace
- * extension's `workspace-server/src/utils/config.ts`. Using the same
- * client_id + redirect_uri lets us piggyback on their published OAuth app
- * (so grants don't hit the 7-day testing-mode expiry).
- */
-const GOOGLE_OAUTH_CLIENT_ID =
-  '338689075775-o75k922vn5fdl18qergr96rp8g63e4d7.apps.googleusercontent.com';
-const GOOGLE_OAUTH_REDIRECT_URI =
-  'https://google-workspace-extension.geminicli.com';
-const GOOGLE_OAUTH_AUTH_BASE = 'https://accounts.google.com/o/oauth2/v2/auth';
-
-/**
- * Default scopes — a broad Workspace set covering common needs. Agents
- * can override at start time via the request body.
- */
-const DEFAULT_GOOGLE_SCOPES = [
-  'openid',
-  'https://www.googleapis.com/auth/userinfo.email',
-  'https://www.googleapis.com/auth/userinfo.profile',
-  'https://www.googleapis.com/auth/gmail.readonly',
-  'https://www.googleapis.com/auth/gmail.send',
-  'https://www.googleapis.com/auth/drive.readonly',
-  'https://www.googleapis.com/auth/drive.file',
-  'https://www.googleapis.com/auth/documents.readonly',
-  'https://www.googleapis.com/auth/calendar.readonly',
-  'https://www.googleapis.com/auth/calendar.events',
-  'https://www.googleapis.com/auth/photoslibrary.readonly',
-];
-
-/**
- * POST /api/credentials/oauth/google/start
- *
- * Build a Google OAuth URL that the caller can send to a remote user (via
- * Slack, email, etc.). The URL opens on the user's device, they sign in,
- * and the gemini-cli cloud function returns a page showing the credentials
- * JSON for the user to copy back.
- *
- * Body: { scopes?: string[] }    (optional override)
- * Returns: { success, data: { authUrl } }
- */
-export async function startGoogleOAuth(
-  req: Request,
-  res: Response,
-  _next: NextFunction,
-): Promise<void> {
-  try {
-    const scopesOverride = (req.body?.scopes as string[] | undefined) ?? undefined;
-    const scopes = scopesOverride && scopesOverride.length > 0
-      ? scopesOverride
-      : DEFAULT_GOOGLE_SCOPES;
-
-    // State payload tells the cloud function to render the manual/copy-paste page
-    // (instead of redirecting to a localhost URI, which this flow doesn't use).
-    const statePayload = { manual: true };
-    const state = Buffer.from(JSON.stringify(statePayload)).toString('base64');
-
-    const params = new URLSearchParams({
-      client_id: GOOGLE_OAUTH_CLIENT_ID,
-      redirect_uri: GOOGLE_OAUTH_REDIRECT_URI,
-      response_type: 'code',
-      scope: scopes.join(' '),
-      state,
-      access_type: 'offline',
-      prompt: 'consent',
-    });
-
-    const authUrl = `${GOOGLE_OAUTH_AUTH_BASE}?${params.toString()}`;
-    res.json({ success: true, data: { authUrl } });
-  } catch (err) {
-    logger.error('startGoogleOAuth failed', { err });
-    res.status(500).json({
-      success: false,
-      error: err instanceof Error ? err.message : 'Failed to build OAuth URL',
-    });
-  }
-}
-
-/**
- * POST /api/credentials/oauth/google/complete
- *
- * Accept the credentials JSON that the user copied from the OAuth success
- * page, verify it, fetch the account email, and save as a new Crewly
- * credential.
- *
- * Body: {
- *   name: string,
- *   credentialsJson: string | object    (the JSON blob the user pasted)
- * }
- * Returns: { success, data: CredentialRegistryEntry }
- */
-export async function completeGoogleOAuth(
-  req: Request,
-  res: Response,
-  _next: NextFunction,
-): Promise<void> {
-  try {
-    const { name, credentialsJson } = (req.body ?? {}) as {
-      name?: string;
-      credentialsJson?: unknown;
-    };
-
-    if (!name || !credentialsJson) {
-      res.status(400).json({
-        success: false,
-        error: 'Missing required fields: name, credentialsJson',
-      });
-      return;
-    }
-
-    // Accept either a parsed object or a raw JSON string
-    let parsed: Record<string, unknown>;
-    if (typeof credentialsJson === 'string') {
-      try {
-        parsed = JSON.parse(credentialsJson);
-      } catch {
-        res.status(400).json({
-          success: false,
-          error:
-            'credentialsJson is not valid JSON. Paste the JSON block from the "Success! Credentials Ready" page exactly as shown.',
-        });
-        return;
-      }
-    } else if (typeof credentialsJson === 'object' && credentialsJson !== null) {
-      parsed = credentialsJson as Record<string, unknown>;
-    } else {
-      res.status(400).json({
-        success: false,
-        error: 'credentialsJson must be a JSON string or object',
-      });
-      return;
-    }
-
-    const accessToken = parsed.access_token as string | undefined;
-    const refreshToken = parsed.refresh_token as string | undefined;
-    const scope = parsed.scope as string | undefined;
-    const tokenType = (parsed.token_type as string | undefined) ?? 'Bearer';
-    const expiryDate = parsed.expiry_date as number | undefined;
-
-    if (!accessToken || !refreshToken) {
-      res.status(400).json({
-        success: false,
-        error:
-          'credentialsJson is missing access_token or refresh_token. Make sure you copied the entire JSON block.',
-      });
-      return;
-    }
-
-    // Fetch account email from Google userinfo (best-effort)
-    let accountEmail: string | undefined;
-    try {
-      const response = await fetch(
-        'https://www.googleapis.com/oauth2/v3/userinfo',
-        { headers: { Authorization: `Bearer ${accessToken}` } },
-      );
-      if (response.ok) {
-        const info = (await response.json()) as { email?: string };
-        accountEmail = info.email;
-      }
-    } catch {
-      // best-effort only — credential is still valid without accountEmail
-    }
-
-    const scopes = typeof scope === 'string' ? scope.split(' ').filter(Boolean) : [];
-
-    const cred = await getCredentialStoreService().addOAuth({
-      name,
-      provider: 'google',
-      helper: 'gemini-cli-workspace',
-      payload: {
-        type: 'google-oauth',
-        accessToken,
-        refreshToken,
-        tokenType: tokenType as 'Bearer',
-        expiresAt: expiryDate ?? Date.now() + 3600_000,
-        scopes,
-        accountEmail,
-        clientId: GOOGLE_OAUTH_CLIENT_ID,
-      },
-    });
-
-    logger.info('Completed headless Google OAuth', {
-      id: cred.id,
-      accountEmail,
-    });
-
-    res.status(201).json({ success: true, data: cred });
-  } catch (err) {
-    logger.error('completeGoogleOAuth failed', { err });
-    res.status(500).json({
-      success: false,
-      error: err instanceof Error ? err.message : 'Failed to save credential',
-    });
-  }
-}
+// Headless OAuth endpoints (startGoogleOAuth, completeGoogleOAuth) live in
+// ./google-oauth.controller.ts — wired into the router below.
